renamed and reorganized components by view/tab
