+++ frontend/src/components/Gallery/Gallery.css
