+++ frontend/src/components/Home/Antivirals.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import toggletip from  './images/toggletip.svg';
-import './AddInitialCases.css';
+import toggletip from  '../images/toggletip.svg';
+import './InitialCases.css';
 
 
 const Antivirals = ({ onSubmit }) => {
