+++ frontend/src/components/Home/DisplayedParameters.css
