+++ frontend/src/components/Home/InitialCases.css
