+++ frontend/src/components/Home/SpreadMapCount.js
