+++ frontend/src/components/Home/SpreadMapPercent.js
@@ -1,25 +1,22 @@
-
-
-// InfectedMap.js
-import React, { useEffect, useState, useRef } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { MapContainer, TileLayer, GeoJSON, useMap } from 'react-leaflet';
 import L, { transformation } from 'leaflet';
 import 'leaflet/dist/leaflet.css';
-import texasOutline from '../data/texasOutline.json';
+import texasOutline from '../../data/texasOutline.json';
 import './Legend.css'; // Ensure you have a CSS file for styling
-import '../fonts/fonts.css'
+import '../../fonts/fonts.css';
 import { IoHomeSharp } from "react-icons/io5";
 
 // Function to determine color based on infected count
-const getColor = (infectedCount) => {
-  return infectedCount > 5000 ? '#800026' :
-    infectedCount > 2000 ? '#BD0026' :
-      infectedCount > 1000 ? '#E31A1C' :
-        infectedCount > 500 ? '#FC4E2A' :
-          infectedCount > 200 ? '#FD8D3C' :
-            infectedCount > 100 ? '#FEB24C' :
-              infectedCount > 50 ? '#FED976' :
-                '#FFEDA0';
+const getColor = (infectedPercent) => {
+  return infectedPercent >= 40 ? "#BD0026" :
+         infectedPercent >= 30 ? "#E31A1B" :
+         infectedPercent >= 20 ? "#FC4F2A" :
+         infectedPercent >= 10 ? "#FD8D3D" :
+         infectedPercent >= 5 ? "#FEB24C" :
+         infectedPercent >= 2.5 ? "#FED976" :
+         infectedPercent >= 1 ? "#FFEDA0" :
+                  "#FFEDA0";
 };
 
 // Function to parse Texas outline GeoJSON
@@ -44,7 +41,6 @@ const ResetButton = () => {
   );
 };
 
-// Component to create a legend for the map
 const Legend = () => {
   const map = useMap();
 
@@ -53,15 +49,16 @@ const Legend = () => {
 
     legend.onAdd = function () {
       const div = L.DomUtil.create('div', 'info legend');
+      // the spacing of percent buckets looks uneven, but closely matches our counts (assuming max of 5000)
       const ranges = [
-        { min: 5000, max: Infinity },
-        { min: 2000, max: 5000 },
-        { min: 1000, max: 2000 },
-        { min: 500, max: 1000 },
-        { min: 200, max: 500 },
-        { min: 100, max: 200 },
-        { min: 50, max: 100 },
-        { min: 0, max: 50 }
+        { min: 40, max: Infinity },
+        { min: 30, max: 40 },
+        { min: 20, max: 30 },
+        { min: 10, max: 20 },
+        { min: 5, max: 10 },
+        { min: 2.5, max: 5 },
+        { min: 1, max: 2.5 },
+        { min: 0, max: 1 },
       ];
       const labels = [];
 
@@ -70,12 +67,19 @@ const Legend = () => {
         labels.push(
           `<div class="legend-item">
             <div class="color-box" style="background:${color};"></div>
-            <span>${range.min}${range.max === Infinity ? '+' : `–${range.max}`}</span>
+            <span>
+                ${range.min === 0 ? `< ${range.max}%` :
+                  range.min === 40 ? `> ${range.min}%` :
+                 `${range.min} -`}
+                ${range.max === 1 ? '' :
+                  range.max === Infinity ? '' :
+                 `${range.max}%`}
+            </span>
           </div>`
         );
       });
 
-      div.innerHTML = `<strong>Infected Count</strong>${labels.join('')}`;
+      div.innerHTML = `<strong>Infected (Percent)</strong>${labels.join('')}`;
       return div;
     };
 
@@ -89,36 +93,35 @@ const Legend = () => {
   return null;
 };
 
-// Main component to render the map
-const InfectedMap = ({ eventData, currentIndex}) => {
+
+const InfectedMapPercent = ({ eventData, currentIndex }) => {
   const [countyData, setCountyData] = useState([]);
   const mapRef = useRef();
 
   useEffect(() => {
     const texasCounties = parseTexasOutline(texasOutline);
-    if (eventData && eventData.length > 0 && typeof currentIndex !== 'undefined') {
-      // Get the event data for the specific day (currentIndex)
+    if (eventData && eventData.length > 0 && typeof currentIndex !== "undefined") {
+      // get event data for the specific day
       const specificDayEventData = eventData.find(event => event.day === currentIndex)?.counties || [];
 
       const data = texasCounties.map(county => {
         const countyEvent = specificDayEventData.find(event => event.fips === county.geoid);
         return {
-          county: county.name,
-          fips: county.geoid,
-          infected: countyEvent ? countyEvent.infected : 0
+        county: county.name,
+        fips: county.geoid,
+        infectedPercent: countyEvent ? countyEvent.infectedPercent: 0
         };
       });
       setCountyData(data);
-      // console.log(`County data loaded for day ${currentIndex}:`, data); // Debug log
     }
   }, [eventData, currentIndex]);
 
-  const onEachCounty = (feature, layer) => {
+const onEachCounty = (feature, layer) => {
     const geoid = feature.properties.geoid;
     const countyInfo = countyData.find(item => item.fips === geoid);
 
     if (countyInfo) {
-      const tooltipContent = `${countyInfo.county}: ${countyInfo.infected} infected`;
+      const tooltipContent = `${countyInfo.county}: ${countyInfo.infectedPercent}% infected`;
       layer.bindTooltip(tooltipContent, {
         permanent: false,
         direction: 'auto',
@@ -142,34 +145,15 @@ const InfectedMap = ({ eventData, currentIndex}) => {
 
   const geoJsonStyle = (feature) => {
     const countyInfo = countyData.find(item => item.fips === feature.properties.geoid);
-    const infectedCount = countyInfo ? countyInfo.infected : 0;
+    const infectedPercentCount = countyInfo ? countyInfo.infectedPercent: 0;
     return {
-      fillColor: getColor(infectedCount),
+      fillColor: getColor(infectedPercentCount),
       weight: 1,
       color: 'black',
       fillOpacity: 0.7
     };
   };
 
-  /*
-  useEffect(() => {
-    if (mapRef.current) {
-      mapRef.current.eachLayer(layer => {
-        if (layer instanceof L.GeoJSON) {
-          layer.clearLayers();
-        }
-      });
-    }
-  }, [eventData]);
-*/
-  /*
-
-  useEffect(() => {
-    if (mapRef.current && outlineLayerRef.current) {
-      outlineLayerRef.current.bringToFront(); // Keep the black outline on top
-    }
-  }, []); // Ensure this only runs once when the map is created
-  */
 
   return (
     <div>
@@ -180,7 +164,7 @@ const InfectedMap = ({ eventData, currentIndex}) => {
         zoom={5.4}
         style={{ height: '35em', backgroundColor: 'transparent'}}
         attributionControl={false}
-        whenCreated={mapInstance => { mapRef.current = mapInstance; }}
+        whenCreated={mapInstance => {mapRef.current=mapInstance; }}
       >
         <GeoJSON
           key={JSON.stringify(countyData)}
@@ -195,4 +179,4 @@ const InfectedMap = ({ eventData, currentIndex}) => {
   );
 };
 
-export default InfectedMap;
+export default InfectedMapPercent;
