+++ frontend/src/components/Home/SpreadTable.css
