+++ frontend/src/components/Home/TimelineSlider.css
