+++ frontend/src/components/Home/TimelineSlider.js
